+++ source/Application/Transactions/MoveIn/MasterDataDelivery/ForwardMeteringPointMasterDataHandler.cs
@@ -20,6 +20,7 @@ using Application.MasterData;
 using Application.OutgoingMessages;
 using Application.OutgoingMessages.Common;
 using Domain.Actors;
+using Domain.Documents;
 using Domain.MasterData.Dictionaries;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.AccountingPointCharacteristics;
