+++ source/Domain/OutgoingMessages/IMessageWriter.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Domain.Documents;
+
 namespace Domain.OutgoingMessages;
 
 /// <summary>
