+++ source/Infrastructure/OutgoingMessages/Common/Xml/MessageWriter.cs
@@ -20,6 +20,7 @@ using System.Threading.Tasks;
 using System.Xml;
 using Application.OutgoingMessages.Common;
 using Application.OutgoingMessages.Common.Xml;
+using Domain.Documents;
 using Domain.OutgoingMessages;
 
 namespace Infrastructure.OutgoingMessages.Common.Xml;
