+++ source/IntegrationTests/Application/Transactions/MoveIn/MasterDataDelivery/SendCustomerMasterDataToGridOperatorTests.cs
@@ -16,6 +16,7 @@ using System.Threading.Tasks;
 using Application.Configuration.DataAccess;
 using Application.Transactions.MoveIn.MasterDataDelivery;
 using Domain.Actors;
+using Domain.Documents;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Domain.Transactions.MoveIn;
